feitos as funcoes
